requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
streamlit==1.29.0
plotly==5.18.0
python-dotenv==1.0.0
fpdf2==2.7.8
qrcode==7.4.2
pillow==10.2.0
